+++ R/shorts-S3.R
@@ -141,7 +141,7 @@ residuals.shorts_model <- function(object, ...) {
 print.shorts_fv_profile <- function(x, ...) {
   cat("Estimated Force-Velocity Profile\n")
   cat("--------------------------------\n")
-  print(unlist(x[-13]))
+  print(unlist(x[-19]))
 }